# contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Athopia client API",
    "version": "1.0.0",
    "description": "Genererad ur athopia-web/lib/api-schemas.ts. Redigera aldrig manuellt."
  },
  "servers": [
    {
      "url": "https://athopia.se"
    }
  ],
  "paths": {
...
    "/api/feed/config": {
      "get": {
        "operationId": "FeedConfigResponse",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "content_types": {
                      "anyOf": [
                        {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "national_interest": {
                      "type": "string",
                      "enum": [
                        "none",
                        "some",
                        "second_team"
                      ]
                    }
                  },
                  "additionalProperties": false,
                  "$schema": "http://json-schema.org/draft-07/schema#"
                }
              }
            }
          },
          "403": {
            "description": "Otillräcklig plan — servern är auktoritativ.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "const": "plan_required"
                    },
                    "feature": {
                      "type": "string"
                    },
                    "requiredPlan": {
                      "type": "string",
                      "enum": [
                        "free",
                        "pro",
                        "elite"
                      ]
                    },
                    "upgradePath": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "code",
                    "feature",
                    "requiredPlan",
                    "upgradePath"
                  ],
                  "additionalProperties": false,
                  "$schema": "http://json-schema.org/draft-07/schema#"
                }
              }
            }
          }
        }
      }
    },
...
  }
}